Add routine to read tables
The file is presumed to have blank lines between tables. Each row of a
table is on its own row. The rows have whitespace, possibly multiple
whitespaces, between rows.

The return is a list of tables. Each table is a list or rows, and each
row is a list of cells.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>