Add initialization time report
Showing the run times for parts 1 and 2 often credits the time to build
the data to part 1 only and not part 2, even though both use the same
data. Removing that from part 1 and into an initializtion report will
equalize the comparisons between the 2 parts.

Is not implemented in solutions prior to 2022 Day 8, but it might be
retrofitted eventually.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>